Fix init_dirs before loging, and preliminary Proxmox 9 Beta support
There was an error where the log dir wasn't created before the script would write to a log, so the script would fail if the folder was missing.

I enabled preliminary Proxmox 9 Beta support. Install works for me, and all the functions seem to work. Installing pve-no-subscription repo will give warnings in apt update, as Proxmox doesn't have trixie packages in there but is using the pve-testing repo instead.